added added combine reducers if need more state store
