+++ src/components/InProgress.js
@@ -48,9 +48,9 @@ class InProgress extends Component {
                     </div>
                   </div>
                   <footer className="card-footer">
-                    <a className="card-footer-item" onClick={this.handleMoveCardLeft.bind(this)} id={card.title}>Move Left</a>
+                    <a className="card-footer-item" onClick={this.handleMoveCardLeft.bind(this)} id={card.title}>In Queue</a>
                     <a className="card-footer-item" onClick={this.handleDelete.bind(this)} id={card.title}>Delete</a>
-                    <a className="card-footer-item" onClick={this.handleMoveCardRight.bind(this)} id={card.title}>Move Right</a>
+                    <a className="card-footer-item" onClick={this.handleMoveCardRight.bind(this)} id={card.title}>Done</a>
                   </footer>
                 </div>
               )
