+++ src/index.js
@@ -3,11 +3,11 @@ import ReactDOM from 'react-dom';
 import { createStore } from "redux";
 import { Provider } from "react-redux";
 import './index.css';
-import cards from "./reducers";
+import reducers from "./reducers";
 import App from './containers/App/App.js';
 import registerServiceWorker from './registerServiceWorker';
 
-const store = createStore(cards);
+const store = createStore(reducers);
 
 ReactDOM.render(
   <Provider store={store}>
